Put LMADs in their own module.
The idea is that the LMAD module is fairly principled and generally
useful, and IxFun is just whatever we put together to serve as index
functions.
