+++ futhark.cabal
@@ -288,6 +288,7 @@ library
       Futhark.IR.Mem
       Futhark.IR.Mem.Interval
       Futhark.IR.Mem.IxFun
+      Futhark.IR.Mem.LMAD
       Futhark.IR.Mem.Simplify
       Futhark.IR.Parse
       Futhark.IR.Pretty
